patientlink-system/auth_service/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
django-cors-headers>=4.2.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
sentry-sdk>=2.14.0
pyotp>=2.9.0
requests>=2.31.0
